// src/types/City.ts
/*
    *  ---------------------------------------------  *
    *  -----  City.ts  --  /src/types/City.ts  -----  *
    *  ---------------------------------------------  *
*/

export interface City {
  id: string;
  name: string;
  country: string;
  admin1: string;
  latitude: number;
  longitude: number;
}